move create extractors to another file
