+++ nwb_conversion_tools/utils/__init__.py
@@ -1,4 +1,5 @@
 from .ephys_writer import export_ecephys_to_nwb, NWBConverter, map_si_object_to_writer
-from .si013nwbephyswriter import SI013NwbEphysWriter, create_si013_example
-from .si090nwbephyswriter import SI090NwbEphysWriter, create_si090_example
+from .si013nwbephyswriter import SI013NwbEphysWriter
+from .si090nwbephyswriter import SI090NwbEphysWriter
 from .neonwbephyswriter import NEONwbEphysWriter
+from .createextractors import create_si090_example, create_si013_example
