+++ nwb_conversion_tools/utils/si013nwbephyswriter.py
@@ -216,87 +216,3 @@ class SI013NwbEphysWriter(BaseSINwbEphysWriter):
         super().add_recording(segment_index=segment_index)
         if self._conversion_ops["write_electrical_series"]:
             self.add_epochs()
-
-
-def create_si013_example(seed):
-    channel_ids = [0, 1, 2, 3]
-    num_channels = 4
-    num_frames = 10000
-    num_ttls = 30
-    sampling_frequency = 30000
-    X = np.random.RandomState(seed=seed).normal(0, 1, (num_channels, num_frames))
-    geom = np.random.RandomState(seed=seed).normal(0, 1, (num_channels, 2))
-    X = (X * 100).astype(int)
-    ttls = np.sort(np.random.permutation(num_frames)[:num_ttls])
-
-    RX = se.NumpyRecordingExtractor(timeseries=X, sampling_frequency=sampling_frequency, geom=geom)
-    RX.set_ttls(ttls)
-    RX.set_channel_locations([0, 0], channel_ids=0)
-    RX.add_epoch("epoch1", 0, 10)
-    RX.add_epoch("epoch2", 10, 20)
-    for i, channel_id in enumerate(RX.get_channel_ids()):
-        RX.set_channel_property(channel_id=channel_id, property_name="shared_channel_prop", value=i)
-
-    RX2 = se.NumpyRecordingExtractor(timeseries=X, sampling_frequency=sampling_frequency, geom=geom)
-    RX2.copy_epochs(RX)
-    times = np.arange(RX.get_num_frames()) / RX.get_sampling_frequency() + 5
-    RX2.set_times(times)
-
-    RX3 = se.NumpyRecordingExtractor(timeseries=X, sampling_frequency=sampling_frequency, geom=geom)
-
-    SX = se.NumpySortingExtractor()
-    SX.set_sampling_frequency(sampling_frequency)
-    spike_times = [200, 300, 400]
-    train1 = np.sort(np.rint(np.random.RandomState(seed=seed).uniform(0, num_frames, spike_times[0])).astype(int))
-    SX.add_unit(unit_id=1, times=train1)
-    SX.add_unit(unit_id=2, times=np.sort(np.random.RandomState(seed=seed).uniform(0, num_frames, spike_times[1])))
-    SX.add_unit(unit_id=3, times=np.sort(np.random.RandomState(seed=seed).uniform(0, num_frames, spike_times[2])))
-    SX.set_unit_property(unit_id=1, property_name="stability", value=80)
-    SX.add_epoch("epoch1", 0, 10)
-    SX.add_epoch("epoch2", 10, 20)
-
-    SX2 = se.NumpySortingExtractor()
-    SX2.set_sampling_frequency(sampling_frequency)
-    spike_times2 = [100, 150, 450]
-    train2 = np.rint(np.random.RandomState(seed=seed).uniform(0, num_frames, spike_times2[0])).astype(int)
-    SX2.add_unit(unit_id=3, times=train2)
-    SX2.add_unit(unit_id=4, times=np.random.RandomState(seed=seed).uniform(0, num_frames, spike_times2[1]))
-    SX2.add_unit(unit_id=5, times=np.random.RandomState(seed=seed).uniform(0, num_frames, spike_times2[2]))
-    SX2.set_unit_property(unit_id=4, property_name="stability", value=80)
-    SX2.set_unit_spike_features(unit_id=3, feature_name="widths", value=np.asarray([3] * spike_times2[0]))
-    SX2.copy_epochs(SX)
-    SX2.copy_times(RX2)
-    for i, unit_id in enumerate(SX2.get_unit_ids()):
-        SX2.set_unit_property(unit_id=unit_id, property_name="shared_unit_prop", value=i)
-        SX2.set_unit_spike_features(
-            unit_id=unit_id, feature_name="shared_unit_feature", value=np.asarray([i] * spike_times2[i])
-        )
-
-    SX3 = se.NumpySortingExtractor()
-    train3 = np.asarray([1, 20, 21, 35, 38, 45, 46, 47])
-    SX3.add_unit(unit_id=0, times=train3)
-    features3 = np.asarray([0, 5, 10, 15, 20, 25, 30, 35])
-    features4 = np.asarray([0, 10, 20, 30])
-    feature4_idx = np.asarray([0, 2, 4, 6])
-    SX3.set_unit_spike_features(unit_id=0, feature_name="dummy", value=features3)
-    SX3.set_unit_spike_features(unit_id=0, feature_name="dummy2", value=features4, indexes=feature4_idx)
-
-    example_info = dict(
-        channel_ids=channel_ids,
-        num_channels=num_channels,
-        num_frames=num_frames,
-        sampling_frequency=sampling_frequency,
-        unit_ids=[1, 2, 3],
-        train1=train1,
-        train2=train2,
-        train3=train3,
-        features3=features3,
-        unit_prop=80,
-        channel_prop=(0, 0),
-        ttls=ttls,
-        epochs_info=((0, 10), (10, 20)),
-        geom=geom,
-        times=times,
-    )
-
-    return (RX, RX2, RX3, SX, SX2, SX3, example_info)
